Release: fast-encrypt@v1.0.0

+++ src/lib/Encryptor.js
@@ -6,10 +6,21 @@ const { now } = require("../utils/functions.js");
 
 class Encryptor {
   static generate_decryptor(hash, format = {}, config = {}) {
-    const { filename = "decrypt.json", pretty = true } = config;
+    const { output, pretty = true } = config;
+    const supported_extension = [".json"];
+    
+    let ext = path.extname(output.filename);
+    if (!supported_extension.includes(ext)) throw new Error(`Unsupported extension for decryptor file expected [${ supported_extension.toString().replace(",", "|") }]`);
+    
+    let path_output = path.join(process.cwd(), output.path);
+    try {
+      if (!fs.existsSync(path_output)) fs.mkdirSync(path_output, { recursive: true });
+    } catch (e) {
+      throw new Error(`can't create path of folders`);
+    }
     
     fs.writeFileSync(
-      path.join(__dirname, filename),
+      path.join(path_output, output.filename),
       JSON.stringify({
         hash: hash,
         format: format,
@@ -17,23 +28,20 @@ class Encryptor {
       }, null, pretty ? 2 : 0)
     );
     
-    console.log(`create file ${ filename }`);
+    console.log(`create file ${ output.filename }`);
   }
   
   static encrypt (text, config = {}) {
     let chars = text.split("");
-    const format = Formatter.create({
-      
-    });
-    
+    let format = Formatter.create();
     let result = chars.map(char => format[char] || " ").join("");
     
-    Encryptor.generate_decryptor(result, format);
+    Encryptor.generate_decryptor(result, format, config);
     return result;
   }
   
   static decrypt (encrypted, formats) {
-    let res = ""
+    let res = "";
     for (let key in formats) {
       let format = formats[key];
       let search = encrypted.search(format);
